pyproject.toml
[tool.poetry]
name = "translate-landofsika"
version = "0.1.0"
description = ""
authors = ["thomas_soutif"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
click = "^8.1.3"
lxml = "^4.9.1"
deepl = "^1.9.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
